# Remove sort behavior from Store.Read RPC, fix Tags invariant violation