--- a/cmd/influxd/launcher/query_test.go
+++ b/cmd/influxd/launcher/query_test.go
@@ -46,17 +46,18 @@ mem,server=b value=45.2`))
 	rawQ := fmt.Sprintf(`from(bucket:"%s")
 	|> range(start:-1m)
 	|> filter(fn: (r) => r._measurement ==  "cpu" and (r._field == "v1" or r._field == "v0"))
+	|> group(columns:["_time", "_value"], mode:"except")
 	`, be.Bucket.Name)
 
 	// Expected keys:
 	//
-	// _measurement=cpu,region=east,server=b,area=z,_field=v1
 	// _measurement=cpu,region=west,server=a,_field=v0
 	// _measurement=cpu,region=west,server=b,_field=v0
+	// _measurement=cpu,region=east,server=b,area=z,_field=v1
 	//
 	results := be.MustExecuteQuery(rawQ)
 	defer results.Done()
-	results.First(t).HasTablesWithCols([]int{5, 4, 4})
+	results.First(t).HasTablesWithCols([]int{4, 4, 5})
 }
 
 // This test initialises a default launcher writes some data,