--- a/storage/reads/merge_test.go
+++ b/storage/reads/merge_test.go
@@ -19,6 +19,115 @@ func newStreamSeries(v ...string) *sliceStreamReader {
 	return newStreamReader(response(frames...))
 }
 
+func TestNewSequenceResultSet(t *testing.T) {
+	tests := []struct {
+		name    string
+		streams []*sliceStreamReader
+		exp     string
+	}{
+		{
+			name: "outer inner",
+			streams: []*sliceStreamReader{
+				newStreamSeries("m0,tag0=val01", "m0,tag0=val02"),
+				newStreamSeries("m0,tag0=val00", "m0,tag0=val03"),
+			},
+			exp: `series: _m=m0,tag0=val01
+  cursor:Float
+series: _m=m0,tag0=val02
+  cursor:Float
+series: _m=m0,tag0=val00
+  cursor:Float
+series: _m=m0,tag0=val03
+  cursor:Float
+`,
+		},
+		{
+			name: "sequential",
+			streams: []*sliceStreamReader{
+				newStreamSeries("m0,tag0=val00", "m0,tag0=val01"),
+				newStreamSeries("m0,tag0=val02", "m0,tag0=val03"),
+			},
+			exp: `series: _m=m0,tag0=val00
+  cursor:Float
+series: _m=m0,tag0=val01
+  cursor:Float
+series: _m=m0,tag0=val02
+  cursor:Float
+series: _m=m0,tag0=val03
+  cursor:Float
+`,
+		},
+		{
+			name: "single resultset",
+			streams: []*sliceStreamReader{
+				newStreamSeries("m0,tag0=val00", "m0,tag0=val01", "m0,tag0=val02", "m0,tag0=val03"),
+			},
+			exp: `series: _m=m0,tag0=val00
+  cursor:Float
+series: _m=m0,tag0=val01
+  cursor:Float
+series: _m=m0,tag0=val02
+  cursor:Float
+series: _m=m0,tag0=val03
+  cursor:Float
+`,
+		},
+		{
+			name: "single series ordered",
+			streams: []*sliceStreamReader{
+				newStreamSeries("m0,tag0=val00"),
+				newStreamSeries("m0,tag0=val01"),
+				newStreamSeries("m0,tag0=val02"),
+				newStreamSeries("m0,tag0=val03"),
+			},
+			exp: `series: _m=m0,tag0=val00
+  cursor:Float
+series: _m=m0,tag0=val01
+  cursor:Float
+series: _m=m0,tag0=val02
+  cursor:Float
+series: _m=m0,tag0=val03
+  cursor:Float
+`,
+		},
+		{
+			name: "single series random order",
+			streams: []*sliceStreamReader{
+				newStreamSeries("m0,tag0=val02"),
+				newStreamSeries("m0,tag0=val03"),
+				newStreamSeries("m0,tag0=val00"),
+				newStreamSeries("m0,tag0=val01"),
+			},
+			exp: `series: _m=m0,tag0=val02
+  cursor:Float
+series: _m=m0,tag0=val03
+  cursor:Float
+series: _m=m0,tag0=val00
+  cursor:Float
+series: _m=m0,tag0=val01
+  cursor:Float
+`,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			rss := make([]reads.ResultSet, len(tt.streams))
+			for i := range tt.streams {
+				rss[i] = reads.NewResultSetStreamReader(tt.streams[i])
+			}
+
+			rs := reads.NewSequenceResultSet(rss)
+			sb := new(strings.Builder)
+			ResultSetToString(sb, rs)
+
+			if got := sb.String(); !cmp.Equal(got, tt.exp) {
+				t.Errorf("unexpected value; -got/+exp\n%s", cmp.Diff(got, tt.exp))
+			}
+		})
+	}
+}
+
 func TestNewMergedResultSet(t *testing.T) {
 	exp := `series: _m=m0,tag0=val00
   cursor:Float