This PR addresses #13581 and comes with some small perf improvements. It will also improve group read performance as it no longer does a redundant sort of the series keys.

### Before

`Store.Read`

```
Consuming data...
Series:   827942
Points:   0
Time:     8163ms
Series/s: 101422.339
Points/s: 0.000
```

`engine.CreateSeriesCursor`

```
Series:   827942
Time:     4612ms
Series/s: 179532.229
```



### After

`Store.Read`

```
Consuming data...
Series:   827942
Points:   0
Time:     5313ms
Series/s: 155844.129
Points/s: 0.000
```

`engine.CreateSeriesCursor`

```
Series:   827942
Time:     732ms
Series/s: 1130639.508
```


Closes #13581